feat: add programmatic API
