+++ package.json
@@ -2,7 +2,7 @@
   "name": "json-autotranslate",
   "description": "Translate a folder of JSON files containing translations into multiple languages.",
   "version": "1.16.1",
-  "main": "index.js",
+  "main": "lib/api.js",
   "license": "MIT",
   "scripts": {
     "start": "ts-node src/index.ts",
